Refactored Company registration to use Layout Controller
